% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data.R
\name{convert_phenos}
\alias{convert_phenos}
\title{Convert phenotypic data to a list_phenos object.}
\usage{
convert_phenos(phenos)
}
\arguments{
\item{phenos}{Matrix of phenotypes (rows observations, columns phenotypic
variables), without missing values.}
}
\value{
An object of class "\code{list_phenos}".
 \item{phenos}{Matrix containing the original phenotypic data.}
 \item{var_err}{Vector containing the sample phenotypic variances.}
}
\description{
This function creates a list_phenos object from a data matrix supplied as
input.
}
\examples{
user_seed <- 123; set.seed(user_seed)
n <- 500; d <- 1000
phenos <- matrix(rnorm(n * d), nrow = n)

list_phenos <- convert_phenos(phenos)

}
\seealso{
\code{\link{convert_snps}}, \code{\link{generate_snps}},
  \code{\link{replicate_real_snps}}, \code{\link{generate_phenos}},
  \code{\link{replicate_real_phenos}}, \code{\link{generate_dependence}}
}
