% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data.R
\name{convert_snps}
\alias{convert_snps}
\title{Convert SNP data to a list_snps object.}
\usage{
convert_snps(snps)
}
\arguments{
\item{snps}{SNP data matrix without missing values. The entries must be 0, 1
or 2.}
}
\value{
An object of class "\code{list_snps}".
 \item{snps}{Matrix containing the original SNP data.}
 \item{vec_maf}{Vector containing the SNP minor allele frequencies.}
}
\description{
This function creates a list_snps object from a SNP data matrix supplied as
input.
}
\examples{
user_seed <- 123; set.seed(user_seed)
n <- 500; p <- 7500
snps <- matrix(sample(0:2, size = n * p, replace = TRUE), nrow = n)

list_snps <- convert_snps(snps)

}
\seealso{
\code{\link{generate_snps}}, \code{\link{replicate_real_snps}},
  \code{\link{convert_phenos}}, \code{\link{generate_phenos}},
  \code{\link{replicate_real_phenos}}, \code{\link{generate_dependence}}
}
