% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoseq_package.R
\docType{package}
\name{echoseq-package}
\alias{echoseq-package}
\title{echoseq: a package for faithful replications and simulations of genetic 
variants, molecular expression levels and other phenotypic data}
\description{
The echoseq package provides functions to emulate molecular quantitative trait
locus data and clinical data associated with genetic variants under
user-specified association patterns. The data can be the result of pure
simulation or can replicate real datasets supplied by the user (which may
replace real data when these cannot be shared for diverse privacy reasons).
The data generation schemes are based on generally accepted principles of
population genetics (Hardy--Weinberg equilibrium, linkage-disequilibrium,
natural selection, pleiotropic control, sparsity assumptions).
}
\section{echoseq functions}{
 generate_dependence, generate_phenos,
generate_snps, replicate_real_phenos, replicate_real_snps.
}

