% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_annotation_driven_data.R
\name{generate_dependence_from_annots}
\alias{generate_dependence_from_annots}
\title{Epigenome-driven QTL mapping}
\usage{
generate_dependence_from_annots(
  n,
  n_loci,
  mean_locus_size,
  p0,
  rho_min_x,
  rho_max_x,
  n_modules,
  mean_module_size,
  rho_min_y,
  rho_max_y,
  r,
  r0,
  prop_act,
  max_tot_pve,
  annots_vs_indep = 1,
  min_dist = 0,
  maf_thres = 0.05,
  max_nb_act_snps_per_locus = 3,
  vec_q = NULL,
  real_snp_mat = NULL,
  real_annot_mat = NULL,
  sd_act_beta = NULL,
  q_pres_annot_loci = NULL,
  bin_annot_freq = 0.05,
  candidate_modules_annots = 1:n_modules,
  tpois_lam_act_annots_mm = 1,
  sd_act_prob = 1,
  sd_err = 1,
  rbeta_sh1_rr = 1,
  n_cpus = 1,
  maxit = 10000,
  module_specific = FALSE
)
}
\description{
Epigenome-driven QTL mapping
}
\examples{
user_seed <- 123; set.seed(user_seed)

# Number of samples
#
n <- 500

# Loci
#
n_loci <- 20
mean_locus_size <- 100
p0 <- 10

# Modules of traits
#
n_modules <- 5
mean_module_size <- 50

# Autocorrelation within loci and equicorrelation within trait modules
#
rho_min_x <- rho_min_y <- 0.5
rho_max_x <- rho_max_y <- 0.9

# Annotations
#
r <- 200
r0 <- 10

# Association pattern
#
prop_act <- 0.1
max_tot_pve <- 0.5

list_assoc <- generate_dependence_from_annots(n, n_loci, mean_locus_size, p0,
                                              rho_min_x, rho_max_x,
                                              n_modules, mean_module_size,
                                              rho_min_y, rho_max_y, r, r0,
                                              prop_act, max_tot_pve)

}
